app/server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-cli
fastapi-cli==0.0.2
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   ollama
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.3
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via pgvector
ollama==0.3.3
    # via -r requirements.in
orjson==3.10.1
    # via fastapi
pgvector==0.2.5
    # via -r requirements.in
postgres==4.0
    # via -r requirements.in
psycopg==3.1.18
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via
    #   postgres
    #   psycopg2-pool
psycopg2-pool==1.2
    # via postgres
pydantic==2.7.0
    # via fastapi
pydantic-core==2.18.1
    # via pydantic
pygments==2.17.2
    # via rich
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
regex==2024.4.16
    # via tiktoken
requests==2.31.0
    # via tiktoken
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tiktoken==0.6.0
    # via -r requirements.in
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   typer
ujson==5.9.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
